Rename UserService.ts to FirebaseUserService.ts to inidcate firebase-base authentication
